예금 상품 진행중 팀원의 기습 푸시로 일단 저장

+++ moneytree_back/src/main/java/com/moneytree_back/domain/DepositAccount.java
@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.UpdateTimestamp;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -29,7 +30,7 @@ public class DepositAccount {
     // @Column(name = "dep") // 멤버는 보류
 
     @Column(name = "deposit_amount", nullable = false, precision = 15, scale = 2)
-    private Double depositAmount;
+    private BigDecimal depositAmount;
 
     @Column(name = "deposit_start_date", nullable = false)
     private LocalDate depositStartDate;
